move chatbox content to source folder
